+++ zero.core/data/market/IoMarketDataModel.cs
@@ -6,34 +6,34 @@ namespace zero.core.data.market
 {
     public class IoMarketDataModel
     {
-        public decimal Change24Hour;
-        public decimal ChangeDay;
-        public decimal ChangePct24Hour;
-        public decimal ChangePctDay;
+        public float Change24Hour;
+        public float ChangeDay;
+        public float ChangePct24Hour;
+        public float ChangePctDay;
         public int Flags;
         public string FromSymbol;
-        public decimal High24Hour;
-        public decimal HighDay;
+        public float High24Hour;
+        public float HighDay;
         public string LastMarket;
         public string LastTradeId;
         public long LastUpdate;
-        public decimal LastVolume;
-        public decimal LastVolumeTo;
-        public decimal Low24Hour;
-        public decimal LowDay;
+        public float LastVolume;
+        public float LastVolumeTo;
+        public float Low24Hour;
+        public float LowDay;
         public string Market;
-        public decimal MktCap;
-        public decimal Open24Hour;
-        public decimal OpenDay;
-        public decimal Price;
-        public decimal Supply;
+        public float MktCap;
+        public float Open24Hour;
+        public float OpenDay;
+        public float Price;
+        public float Supply;
         public string ToSymbol;
-        public decimal TotalVolume24h;
-        public decimal TotalVolume24hTo;
+        public float TotalVolume24h;
+        public float TotalVolume24hTo;
         public string Type;
-        public decimal Volume24Hour;
-        public decimal Volume24HourTo;
-        public decimal VolumeDay;
-        public decimal VolumeDatTo;
+        public float Volume24Hour;
+        public float Volume24HourTo;
+        public float VolumeDay;
+        public float VolumeDatTo;
     }
 }
